fix: move views to views folder
