+++ lib/commands/open.js
@@ -1,6 +1,6 @@
 'use babel'
 
-import SelectSavedWorkspaceView from './../select-saved-workspace-view'
+import SelectSavedWorkspaceView from './../views/select-saved-workspace-view'
 import Workspace from './../models/workspace'
 
 const WORKSPACES_PATH = 'save-workspace-Workspaces'
