+++ lib/commands/save.js
@@ -1,6 +1,6 @@
 'use babel'
 
-import InputWorkspaceNameView from './../input-workspace-name-view'
+import InputWorkspaceNameView from './../views/input-workspace-name-view'
 import Workspace from './../models/workspace'
 
 const WORKSPACES_PATH = 'save-workspace-Workspaces'
