+++ lib/views/input-workspace-name-view.js
@@ -1,7 +1,7 @@
 'use babel'
 
 import { SelectListView, TextEditorView } from 'atom-space-pen-views'
-import * as listViewFormatter from './utils/listViewFormatter'
+import * as listViewFormatter from '../utils/listViewFormatter'
 
 class InputWorkspaceNameView extends SelectListView {
