+++ lib/views/select-saved-workspace-view.js
@@ -1,7 +1,7 @@
 'use babel'
 
 import { SelectListView } from 'atom-space-pen-views'
-import * as listViewFormatter from './utils/listViewFormatter'
+import * as listViewFormatter from '../utils/listViewFormatter'
 
 class SelectSavedWorkspaceView extends SelectListView {
